requirements.txt

albumentations
dlib
einops
gradio
imageio
matplotlib
more_itertools
numpy
omegaconf
Pillow
pudb
pytorch_lightning
PyYAML
requests
scikit_image
scipy
setuptools
streamlit
torch
torchvision
tqdm
transformers
typing_extensions

